maltoolbox/lexer_parser/requirements.txt
antlr4-tools
antlr4-python3-runtime
